Implement adding part, refactor styles

+++ cms/src/pages/courses/[id]/overview.tsx
@@ -5,7 +5,7 @@ import Link from "next/link"
 import useQueryParameter from "../../../hooks/useQueryParameter"
 import { useQuery } from "react-query"
 import { dontRenderUntilQueryParametersReady } from "../../../utils/dontRenderUntilQueryParametersReady"
-import { Typography, Button, Grid } from "@material-ui/core"
+import { Typography, Grid, Button } from "@material-ui/core"
 import NewPageForm from "../../../components/forms/NewPageForm"
 import { CoursePart, Page } from "../../../services/services.types"
 import { deletePage } from "../../../services/backend/pages"
@@ -13,6 +13,29 @@ import { fetchCourseStructure } from "../../../services/backend/courses"
 import { normalWidthCenteredComponentStyles } from "../../../styles/componentStyles"
 import { css } from "@emotion/css"
 import NewPartForm from "../../../components/forms/NewPartForm"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
+import { faPen, faTrash } from "@fortawesome/free-solid-svg-icons"
+
+import styled from "@emotion/styled"
+import { useRouter } from "next/router"
+
+const DeleteButton = styled.button`
+  color: red;
+  border: 0;
+
+  &:hover {
+    cursor: pointer;
+  }
+`
+
+const EditButton = styled.button`
+  color: blue;
+  border: 0;
+
+  &:hover {
+    cursor: pointer;
+  }
+`
 
 const CoursePages: React.FC<unknown> = () => {
   const id = useQueryParameter("id")
@@ -21,6 +44,7 @@ const CoursePages: React.FC<unknown> = () => {
   )
   const [showNewPageForm, setShowNewPageForm] = useState(false)
   const [showNewPartForm, setShowNewPartForm] = useState(false)
+  const router = useRouter()
 
   if (error) {
     return <div>Error overview.</div>
@@ -30,17 +54,21 @@ const CoursePages: React.FC<unknown> = () => {
     return <div>Loading...</div>
   }
 
-  const handleShowingPageForm = () => {
-    setShowNewPageForm(!showNewPageForm)
-  }
-
-  const handleDelete = async (pageId: string) => {
-    await deletePage(pageId)
-    refetch()
+  const handleDeleteTopLevelPage = async (pageId: string, name: string) => {
+    const result = confirm(`Want to delete ${name}?`)
+    if (result) {
+      await deletePage(pageId)
+      refetch()
+    }
   }
 
   const handleCreateTopLevelPage = () => {
     setShowNewPageForm(!showNewPageForm)
+    refetch()
+  }
+
+  const handleCreatePart = () => {
+    setShowNewPartForm(!showNewPartForm)
     refetch()
   }
 
@@ -55,20 +83,23 @@ const CoursePages: React.FC<unknown> = () => {
           Course overview for <b>{data.course.name}</b>
         </Typography>
         <Grid key={data.course.id} style={{ margin: "0.5em 0" }} container spacing={1}>
-          <Grid item xs={3}>
+          <Grid item xs={4}>
             <h3>Path</h3>
           </Grid>
           <Grid item xs={6}>
             <h3>Title</h3>
           </Grid>
-          <Grid item xs={3}>
-            &nbsp;
+          <Grid item xs={1}>
+            <h3>Edit</h3>
+          </Grid>
+          <Grid item xs={1}>
+            <h3>Delete</h3>
           </Grid>
           {data.pages
             .filter((page) => !page.deleted)
             .map((page: Page) => (
               <>
-                <Grid item xs={3}>
+                <Grid item xs={4}>
                   {page.url_path}
                 </Grid>
                 <Grid item xs={6}>
@@ -76,23 +107,26 @@ const CoursePages: React.FC<unknown> = () => {
                     {page.title}
                   </Link>
                 </Grid>
-                <Grid item xs={3}>
-                  <Button
-                    onClick={() => handleDelete(page.id)}
-                    variant="outlined"
-                    color="secondary"
-                  >
-                    Delete
-                  </Button>
+                <Grid item xs={1}>
+                  <EditButton onClick={() => router.push(`/pages/${page.id}`)}>
+                    <FontAwesomeIcon icon={faPen} size="lg" />
+                  </EditButton>
+                </Grid>
+                <Grid item xs={1}>
+                  <DeleteButton onClick={() => handleDeleteTopLevelPage(page.id, page.title)}>
+                    <FontAwesomeIcon icon={faTrash} size="lg" />
+                  </DeleteButton>
                 </Grid>
               </>
             ))}
         </Grid>
-        {!showNewPageForm && <Button onClick={handleShowingPageForm}>New top level page</Button>}
+        {!showNewPageForm && (
+          <Button onClick={() => setShowNewPageForm(!showNewPageForm)}>New top level page</Button>
+        )}
         {showNewPageForm && (
           <div>
+            <Button onClick={() => setShowNewPageForm(!showNewPageForm)}>Hide</Button>
             <NewPageForm courseId={id} onSubmitForm={handleCreateTopLevelPage} />
-            <Button onClick={handleShowingPageForm}>Hide</Button>
           </div>
         )}
         <div>
@@ -103,11 +137,13 @@ const CoursePages: React.FC<unknown> = () => {
                 <p>{part.name}</p>
               </div>
             ))}
-          {!showNewPartForm && <Button>Add new part</Button>}
+          {!showNewPartForm && (
+            <Button onClick={() => setShowNewPartForm(!showNewPartForm)}>Add new part</Button>
+          )}
           {showNewPartForm && (
             <div>
-              <NewPartForm courseId={id} />
-              <Button>Hide</Button>
+              <Button onClick={() => setShowNewPartForm(!showNewPartForm)}>Hide</Button>
+              <NewPartForm courseId={id} onSubmitForm={handleCreatePart} />
             </div>
           )}
         </div>
